Fix characteristic string check
testCharacteristic("CurrentRelativeHumidity") returns false
testCharacteristic("Current Relative Humidity") returns true

Documentation states to use CurrentRelativeHumidity as the
Characteristic string value for both notifications and mqtt messages.
Best to keep that consistent as it has been historically like that.

This essentially reverts the change to this section which happened in
commit 9fb070699170e88fc418186fd316a99c8ac49661 "Adding support for
MQTT"